Regain lost performance by testing debug messages
Several dprint() calls are invoking Data::Dumper to dump large
data structures, which is an expensive operation. These dumps
were being created in the arguments to the dprint() call regardless
of whether dprint() was actually enabled.

Solution: add an if test to ensure that dprint() is enabled before
constructing its arguments.