Convert instruction labels from strings to integers
One of Plang's original design goals was to have its parser
and compiler be easy to read at debug. Towards that end,
most values are plain strings. This makes it trivial to
dump the AST, etc, without any remapping or prettifying.

Unfortunately, this can be slow. Though the performance
so far has been adequate and acceptable, for future
performance considerations, I have decided that the
ease and convenience of readable strings in the AST
is now a lesser goal to performant and efficient
interpretation.

Converting the strings to integers allows us to use an
array as an extremely efficient look-up table to dispatch
instructions to their implementations. The giant and
horrible if/else chain of string comparisions is no more!